view ct file

+++ apps/simulator/src/c_position.rs
@@ -1,7 +1,7 @@
 use bevy_ecs::prelude::*;
 #[derive(Component, Debug)]
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
-    pub rotation: f32,
+    pub x: f64,
+    pub y: f64,
+    pub rotation: f64,
 }
