+++ apps/simulator/src/c_velocity.rs
@@ -1,11 +1,11 @@
 use bevy_ecs::prelude::*;
 #[derive(Component)]
 pub struct Velocity {
-    pub velocity: f32,
+    pub velocity: f64,
 }
 #[derive(Component)]
 pub struct TankVelocity {
-    pub angular_velocity: f32,
-    pub gun_angular_velocity: f32,
-    pub radar_angular_velocity: f32,
+    pub angular_velocity: f64,
+    pub gun_angular_velocity: f64,
+    pub radar_angular_velocity: f64,
 }
